feat(i2c): add support for 10-bit I2C addresses
This branch adds support for the 10-bit I2C address mode. This was
pretty trivial to add but I had punted on it in PR #103 in the interest
of getting the minimum necessary I2C stuff working. I couldn't easily
test this because I don't have any I2C devices on hand that require the
10-bit address mode, but it should work! :)

Closes #143