fix: remove use of `tracing` v0.2
Currently, we use `tracing` v0.2 on some platforms, and v0.1 on others.
This was motivated by `tracing` v0.1 requiring a single allocation
performed using `liballoc` to allocate an `Arc` when setting the default
subscriber, which is not needed in `tracing` v0.2. This was an issue
when mnemOS did not use `liballoc`. However, we now use `liballoc` on
all platforms, so we can completely cut over from the unreleased
`tracing` v0.2 to released `tracing` v0.1 everywhere.

This branch makes that change, rendering the complex feature flagging
situation previously used to switch between the released and unreleased
`tracing` versions unnecessary. This should fix the build in the root
workspace without feature flags.

We still feature flag the automatic spawning of the serial port
subscriber, so that Melpomene and Pomelo can use their own native
subscribers instead. We may want to change this so that the platform
impl is responsible for setting the global default subscriber, instead,
but this works for now.

Fixes #142