feat(esp32-c3): add stub esp32-c3 project (#193)
This branch adds a mostly empty ESP32-C3 project in `platforms/`, and
adds Just recipes for building and flashing the ESP32-C3. Currently, the
C3 platform crate just prints "hello world"; it does *not* actually run
a working MnemOS kernel. Actual MnemOS bringup on top of the stub
project will be added in a follow-up PR.