RFC: WiFi Buddy Interface Design
This commit starts a new RFC for designing the hardware and software
interfaces between MnemOS on the CPU and MnemOS on the WiFi Buddy. I've
started by adding a summary of the background and a list of design
questions that need to be answered as we determine how the WiFi Buddy
and the CPU will communicate. This is intended as a jumping-off point
for future conversations, and (eventually) to document the design.

@jamesmunns, I'd love your input on this --- feel free to add to the RFC
doc directly, or leave comments on the PR, whatever you prefer!