feat(i2c_puppet): pass in a `WaitCell` for IRQs
This commit updates the `i2c_puppet` driver to allow passing in an
optional `WaitCell` via the `I2cPuppetSettings` struct. This `WaitCell`
can be woken whenever the `i2c_puppet` IRQ line is asserted, to allow
interrupt-driven operation. If no `WaitCell` is provided, the driver
will run in polling-only mode, the current behavior.

We will still poll periodically when the poll interval elapses
regardless of whether a `WaitCell` is provided. This way, we can ensure
that the keyboard is always polled even if an interrupt request is
occasionally missed. Future changes may allow disabling the periodic
polling mode.