feat(mq-pro): wake `i2c_puppet` driver on IRQs (#201)
Depends on #200

This branch changes the MQ Pro board binary to configure an interrupt on
the `i2c_puppet` IRQ pin, and wake the driver when that interrupt fires.

I did *not* implement a general-purpose abstraction for GPIO interrupts
in this branch (see #159). Therefore, this is all hard-coded and
installing this interrupt handler will prevent other code from
registering interrupts for any other PB7 pin. However, I figured that
"perfect is the enemy of working", and, rather than designing and
implementing a more generic GPIO interrupt abstraction in this branch, I
just wanted to get something that worked for the keyboard ISR
specifically. We can build a nice abstraction for interrupts on GPIO in
the future.