chore(esp32c3): add `core` crate in main workspace
This change moves most of the driver and initialization code currently
in `platforms/esp32c3-buddy` into a new `platforms/esp32c3/core` crate,
which is part of the main workspace, with only the actual bins built for
the target binaries in their own workspace.

This allows most of the driver code to participate in stuff like `cargo
check` and `cargo fmt` runs in the main workspace, which should make our
lives a bit easier. It was necessary to ensure that incompatible
features of the `critical-section` crate were not enabled in the main
workspace, so I did some tweaking of the Cargo features to ensure that
we weren't enabling an incompatible configuration in core crates.