feat(hw): Add mnemos shield (#163)
This PR adds the CAD design files for the second version of the mnemos
shield. This design matches the version of the boards that were built
for @hawkw.

The original revision had the center row of headers with 1.2mm PTH,
which were too large to be functional for the pogo pins. This history is
not included (because it didn't fully work).

Gerber exports, suitable for ordering from JLC or other board fabs will
be added as a github release after this has been merged.