chore: make VS Code rust-analyzer run `just clippy`
This commit changes the VS Code settings so that rust-analyzer will run
the `just clippy` command instead of `cargo clippy` when running checks.
Because the `just clippy` command will check all crates in the
workspace, not just the default members, this means VS Code users will
get clippy warnings/errors for *all* crates in the workspace, even the
ones that can't be checked as part of a global `cargo clippy
--workspace` command. This should substantially improve the dev
experience for VS Code users.

This change required adding an optional trailing `ARGS` argument to the
`just clippy` recipe so that additional args can be passed to Clippy (in
this case `--message-format=json --quiet --all-targets`, to mimic the
way `rust-analyzer` normally invokes `cargo check`).