docs: add top-level docs to workspace page
Oranda v0.3.1 now supports the ability to add a top-level Markdown file
to the workspace page, before the list of workspace components. This
branch updates the docs site to use the new version of Oranda, and adds
this top-level docs file so that our main landing page actually says
what MnemOS is.