chore: fix mdbook build in netlify
It turns out that Netlify doesn't check out the repo to a directory
named `mnemos` --- instead, it checks it out to `/opt/build/repo`. This,
of course, means that the check in `rfc2book.py` that we are being run
in a path that ends with either `book` or `mnemos` fails when building
the netlify site.

This commit removes that check to hopefully debreak the docs site build.