chore: nicer out of the box Pomelo dev environment (#254)
This branch modifies the provided dev environment to make working on
Pomelo easier for users who don't have the requisite dev dependencies.
In particular:

- the `wasm32-unknown-unknown` target is added to the
`rust-toolchain.toml` so that rustup will ensure that the target is
installed, if it is not already present.
- a `trunk` recipe is added to the `justfile`, for running `trunk`
commands and installing it if it is not present
- `trunk` is added to the Nix flake