feat(kernel): add listeners to services (#258)
Currently, when a task (or userspace) wishes to connect to a
`RegisteredDriver`, it performs a registry lookup, which returns a
handle for sending requests to that service. These handles are
essentially producers for a single `KChannel` which is associated with
the service when it's registered, and the service can use this channel's
`KConsumer` to receive a stream of incoming requests.

This approach works fine, but it limits what a service can do with its
incoming request stream. Because all requests arrive on the same
`KConsumer`, the server may not spawn a separate task to handle requests
from each client. Similarly, it cannot route requests to different
worker tasks based on which resource the request is associated with.

This branch changes the way clients communicate with servers by
introducing a _handshake_ to this process. Rather than registering a
service by giving the registry a `KProducer` for a single request
channel, the registry now owns a producer for a channel of _incoming
connections_. These connections contain a `Hello` message sent by the
client, of a type defined by the `RegisteredDriver` implementation for
the service. This may be used by the service to determine what resource
the client wants to access, or otherwise route connections to different
tasks. This handshake is completed by the service responding to the
incoming connection with a `KProducer` for a request channel, which is
then returned to the client.

The returned `KProducer` _may_ still point to a single global request
channel for that service, allowing us to continue to implement the
existing behavior. However, the service may alternatively spawn a new
task for that connection and create a new channel for that worker. Or,
it may return a different `KProducer` for a different request channel
depending on the `Hello` message, allowing the server to have separate
worker tasks for different resources managed by the same service, and
route the connection to one of those workers depending on which resource
was requested in the `Hello` message.

An adapter, `into_request_stream`, is provided, to make it easier for
services which want the current behavior of one global stream of
requests to have that. Currently, all the existing services use this,
but in the future, many of our existing services can be rewritten to use
the new handshake mechanism more effectively. I'd also like to do some
additional refactoring, but I wanted to get a working PR up first.