fix(esp32c3): fix missing heap link section (#272)
In `esp32c3-hal` v0.12.x, the linker script no longer includes a heap
region, and the heap is instead intended to be defined as a `static mut`
in Rust-land. This meant that after updating from v0.11.0 to v0.12.0,
the `mnemos-esp32c3-buddy` binaries fail to link, because they reference
the no-longer-existing `_heap_start` symbol. We didn't catch this on CI
because we never actually build binaries, just `cargo check`, so
link-time issues were not being caught.

This commit fixes that by using a Rust static for the ESP32C3 heap.