feat(crowtty): more readable tracing `message`s
This commit makes a very small change to how the crowtty traceproto
output formats an event's `message` field. The message is now printed
without the `message=` prefix, shortening the log line, and it's now
bolded so the message stands out from the other fields and metadata. I
think this is a little more readable.