fix(d1): put back missing `serial-trace` feature (#278)
Looks like the feature flag was lost in #271, so the D1 binaries no
longer emit traceproto output. I believe this happened when combining
the `mnemos-d1` crate (which used to enable the feature) with the
`mnemos-d1-core` crate (which didn't enable the feature).

This commit puts it back. :)

I've also modified the Just recipes for building D1 images to allow
passing additional args to Cargo. This way, users of the Just recipes
can now pass `--no-default-features --features ...` to pick which
features are enabled when building. This was necessary for testing what
happened when the board config had a `serial-trace` section but the
feature flag was disabled. But, passing extra cargo args may also be
useful in other situations.