chore(justfile): also run doctests
Because we're using `cargo nextest`, which can't easily learn about
doctests, the `just test` command currently only runs normal unit and
integration tests. This means that we have a bunch of doctests which
were silently not being run. Whoopsie!

This commit adds a `cargo test --doc` invocation to `just test` as well,
so that doctests are also run.