feat(D1): `mycelium_bitfield` DMAC descriptors
Currently, the DMAC descriptor configuration fields are bit-packed every
time a descriptor is constructed. However, the drivers which use DMA
always construct descriptors with the same configuration bitfields --
only the addresses and lengths change. We can make this more efficient
by constructing the config field once and copying it into each new
descriptor.

This branch replaces the existing `DescriptorConfig` struct with a
`DescriptorBuilder`, which is basically the same thing except it
actually constructs the bitfield as a single u32. I used
`mycelium-bitfield` to do this --- mostly for fun, but it's maybe a bit
more readable than the previous code? I dunno.