fix(mn): use correct "_any-deps" feature name (#286)
Manganese uses a "private" feature flag called "_any-deps" to determine
whether any bindep dependency feature flag is enabled. Unfortunately,
some parts of the code mistakenly use the name "_any_deps" (with a
second "_" instead of a "-") to refer to this feature, which is
incorrect. This results in Manganese always emitting a warning that no
deps were installed, even when this is not actually the case.
Surprisingly, Rust doesn't warn you when you reference a feature flag
that doesn't actually exist, which means we never caught this issue in
development.

This branch fixes that by using the correct name for the feature flag
everywhere.

Fixes #285