feat(kernel): initialize tracing subscriber earlier (#288)
This changes the `tracing` initialization so that the subscriber is
registered as the default *without* waiting for SerMux to actually come
up. This way, we can trace stuff that happens before the serial driver
starts: anything we record gets buffered, and then is sent over serial once
serial comes up. This lets us trace more of the boot process.