feat(bitslab): add capacity limit to index allocs
Currently, an `IndexAlloc{N}` from `mnemos-bitslab` will always allocate
up to the number of bits in the numeric type used as the bitmap. In some
cases, it may be desirable to allocate fewer values. Therefore, this
branch adds support for setting an upper bound on the number of indices,
which may be lower than the available bits.