chore(CI): temporarily pin `cargo-nextest` to v0.9.61 (#300)
v0.9.63 of `cargo-nextest` introduced a regression
(https://github.com/guppy-rs/guppy/issues/157#issuecomment-1830490526)
that breaks Nextest when a workspace includes a package with binary
dependencies. This broke our CI builds.

This PR temporarily pins the `cargo-nextest` version on CI to v0.9.61
(as v0.9.62 was yanked). Hopefully when this is fixed upstream, we can
resume using the latest version.