chore(ci): run `rustup toolchain` *after* checkout (#315)
Currently, the `rustup show` commands run on CI don't actually show the
real toolchain that will be used, because they run *before* we actually
check out the repo. Because we haven't checked out the repo, the
`rust-toolchain.toml` file...isn't there, so we don't have a configured
toolchain.

This commit fixes that. We now run `rustup show` *after* checking out
the repo, so the `rustup show` command shows the toolchain that CI will
actually use.