tools: rewrite x86_64 build system (#334)
So I rewrote the x86_64 kernel's build system again.

This commit rewrites the builder for the x86_64 kernel to be a cargo
custom runner, rather than a build script in a normal crate. This way,
we can use

```console
$ cargo run --package mnemos-x86_64 --target-arch x86_64-unknown-none
```

to build the x86 kernel normally using cargo, outputting a path that's
then passed into the custom runner binary by Cargo. This way, we don't
build the whole kernel inside a build script, so we get normal cargo
output from the kernel build, and we don't need to deal with the fact
that we have to rerun the build script whenever the kernel or any of its
dependencies changes. Instead, the kernel is built by a normal cargo
invocation, which is like infinitely better in literally every possible
way. The custom runner can also launch QEMU, so you can now `cargo run`
the x86 kernel and IT JUST WORKS, RIGHT OUT OF THE BOX.[^1]

The one weird thing about this approach is that now, you build the
kernel by running `cargo run` with a `build` subcommand that tells it to
not actually build the kernel, instead of `cargo build` like a normal
person. This is, admittedly, weird; but it's so much less bad than the
previous thing where you had to run `cargo clean` every time you changed
the kernel if you didn't want to get a stale build artifact (see #332),
which was *incredibly life-ruining* and made developing the x86 kernel a
horrific nightmare of infinite pain. And, most of the time you'll build
it using the Justfile `just run-x86` and `just build-x86` recipes so you
don't have to face the horrible truth that you're using `cargo run` to
build something without running it. So, whatever.

Closes #332

[^1]: Provided that, like, you have `qemu-system-x86_64` on your
    computer, and stuff.