tools: separate crowtty bin from libuv dep, fix CI (#342)
The `crowtty` CLI requires `libuv` on Linux systems in order to interact
with serial devices. This is why the x86_64 build is now failing on CI,
since the bootimage builder depends on `crowtty` in order to run it on
the QEMU virtual serial port, but building crowtty requires `libuv`.
See:
https://github.com/tosc-rs/mnemos/issues/341#issuecomment-2395519851

This commit separates the `crowtty` binary from the `libcrowtty`
library, so that only the CLI binary depends on `libuv`. Hopefully this
fixes the x86_64 CI build.

Fixes #341